Slide Jquery
Slide-JQuery
============

Hey guys, this code is a slide of images
without using plugins created from scratch with jquery library.

Using
it is easy just need to add the files to the appropriate folders and
change the address of the pages in the "style.css" file.

If you have
any questions can consult troka-mc@hotmail.com sending me an email.

I
hope you enjoy this and it is of great help in their web projects

Here
I leave a video of how it works:
https://www.youtube.com/watch?v=UbTb31zcRzc